componentized PresentationTab
